Add BeagleBoneBlack support
To support another board the gpio.cc and gpio.h have been renamed
to rpi_gpio.cc and rpi_gpio.h. And the equivalent files have been
created for the BeagleBoneBlack: bbb_gpio.cc and bbb_gpio.h

The Makefile needs to be edited to select which board to build for.